+++ cfripper/rules/wildcard_principals.py
@@ -1,12 +1,25 @@
-__all__ = ["GenericWildcardPrincipalRule", "PartialWildcardPrincipalRule", "FullWildcardPrincipalRule"]
+__all__ = [
+    "GenericWildcardPrincipalRule",
+    "PartialWildcardPrincipalRule",
+    "FullWildcardPrincipalRule",
+    "GenericResourceWildcardPrincipalRule",
+    "GenericResourcePartialWildcardPrincipalRule",
+    "GenericResourceFullWildcardPrincipalRule",
+]
 import logging
 import re
 from typing import Dict, Optional
 
 from pycfmodel.model.cf_model import CFModel
+from pycfmodel.model.resources.iam_managed_policy import IAMManagedPolicy
+from pycfmodel.model.resources.iam_policy import IAMPolicy
 from pycfmodel.model.resources.iam_role import IAMRole
+from pycfmodel.model.resources.iam_user import IAMUser
 from pycfmodel.model.resources.kms_key import KMSKey
 from pycfmodel.model.resources.properties.policy_document import PolicyDocument
+from pycfmodel.model.resources.s3_bucket_policy import S3BucketPolicy
+from pycfmodel.model.resources.sns_topic_policy import SNSTopicPolicy
+from pycfmodel.model.resources.sqs_queue_policy import SQSQueuePolicy
 
 from cfripper.config.regex import REGEX_FULL_WILDCARD_PRINCIPAL, REGEX_PARTIAL_WILDCARD_PRINCIPAL
 from cfripper.model.enums import RuleGranularity, RuleRisk
@@ -19,6 +32,7 @@ logger = logging.getLogger(__file__)
 class GenericWildcardPrincipalRule(PrincipalCheckingRule):
     """
     Checks for any wildcard principal defined in any statement.
+    Only for: IAMManagedPolicy, IAMPolicy, S3BucketPolicy, SNSTopicPolicy and SQSQueuePolicy
     To be inherited into more precise rules.
     Ignores KMS Keys, since they have `KMSKeyWildcardPrincipalRule`.
     For IAM Roles, it also checks `AssumeRolePolicyDocument`.
@@ -52,14 +66,14 @@ class GenericWildcardPrincipalRule(PrincipalCheckingRule):
     def invoke(self, cfmodel: CFModel, extras: Optional[Dict] = None) -> Result:
         result = Result()
         for logical_id, resource in cfmodel.Resources.items():
-            if isinstance(resource, KMSKey):
-                # Ignoring KMSKey because there's already a rule for it `KMSKeyWildcardPrincipalRule`
-                continue
-            if isinstance(resource, IAMRole):
-                # Checking the `AssumeRolePolicyDocument` for IAM Roles
-                self.check_for_wildcards(result, logical_id, resource.Properties.AssumeRolePolicyDocument, extras)
-            for policy in resource.policy_documents:
-                self.check_for_wildcards(result, logical_id, policy.policy_document, extras)
+            if isinstance(resource, (IAMManagedPolicy, IAMPolicy, S3BucketPolicy, SNSTopicPolicy, SQSQueuePolicy)):
+                self.check_for_wildcards(result, logical_id, resource.Properties.PolicyDocument, extras)
+            elif isinstance(resource, (IAMRole, IAMUser)):
+                if isinstance(resource, IAMRole):
+                    self.check_for_wildcards(result, logical_id, resource.Properties.AssumeRolePolicyDocument, extras)
+                if resource.Properties and resource.Properties.Policies:
+                    for policy in resource.Properties.Policies:
+                        self.check_for_wildcards(result, logical_id, policy.PolicyDocument, extras)
 
         return result
 
@@ -130,6 +144,105 @@ class PartialWildcardPrincipalRule(GenericWildcardPrincipalRule):
 
 
 class FullWildcardPrincipalRule(GenericWildcardPrincipalRule):
+    """
+    Checks for any wildcard principal defined in any statement.
+
+    Risk:
+        It might allow other AWS identities to escalate privileges.
+
+    Fix:
+        Where possible, restrict the access to only the required resources.
+        For example, instead of `Principal: "*"`, include a list of the roles that need access.
+
+    Filters context:
+        | Parameter   | Type               | Description                                                    |
+        |:-----------:|:------------------:|:--------------------------------------------------------------:|
+        |`config`     | `str`              | `config` variable available inside the rule                    |
+        |`extras`     | `str`              | `extras` variable available inside the rule                    |
+        |`logical_id` | `str`              | ID used in CloudFormation to refer the resource being analysed |
+        |`resource`   | `S3BucketPolicy`   | Resource that is being addressed                               |
+        |`statement`  | `Statement`        | Statement being checked found in the Resource                  |
+        |`principal`  | `str`              | AWS Principal being checked found in the statement             |
+        |`account_id` | `str`              | Account ID found in the principal                              |
+    """
+
+    RISK_VALUE = RuleRisk.HIGH
+
+
+class GenericResourceWildcardPrincipalRule(GenericWildcardPrincipalRule):
+    """
+    Checks for any wildcard principal defined in any statement for any type of resource.
+    To be inherited into more precise rules.
+    Ignores KMS Keys, since they have `KMSKeyWildcardPrincipalRule`.
+    For IAM Roles, it also checks `AssumeRolePolicyDocument`.
+
+    Risk:
+        It might allow other AWS identities to escalate privileges.
+
+    Fix:
+        Where possible, restrict the access to only the required resources.
+        For example, instead of `Principal: "*"`, include a list of the roles that need access.
+
+    Filters context:
+        | Parameter   | Type               | Description                                                    |
+        |:-----------:|:------------------:|:--------------------------------------------------------------:|
+        |`config`     | `str`              | `config` variable available inside the rule                    |
+        |`extras`     | `str`              | `extras` variable available inside the rule                    |
+        |`logical_id` | `str`              | ID used in CloudFormation to refer the resource being analysed |
+        |`resource`   | `S3BucketPolicy`   | Resource that is being addressed                               |
+        |`statement`  | `Statement`        | Statement being checked found in the Resource                  |
+        |`principal`  | `str`              | AWS Principal being checked found in the statement             |
+        |`account_id` | `str`              | Account ID found in the principal                              |
+    """
+
+    def invoke(self, cfmodel: CFModel, extras: Optional[Dict] = None) -> Result:
+        result = Result()
+        for logical_id, resource in cfmodel.Resources.items():
+            if isinstance(resource, KMSKey):
+                # Ignoring KMSKey because there's already a rule for it `KMSKeyWildcardPrincipalRule`
+                continue
+            if isinstance(resource, IAMRole):
+                # Checking the `AssumeRolePolicyDocument` for IAM Roles
+                self.check_for_wildcards(result, logical_id, resource.Properties.AssumeRolePolicyDocument, extras)
+            for policy in resource.policy_documents:
+                self.check_for_wildcards(result, logical_id, policy.policy_document, extras)
+
+        return result
+
+
+class GenericResourcePartialWildcardPrincipalRule(GenericResourceWildcardPrincipalRule):
+    """
+    Checks for any wildcard or account-wide principals defined in any statements. This rule will flag
+    as non-compliant any principals where `root` or `*` are included at the end of the value, for
+    example, `arn:aws:iam:12345:12345*`.
+
+    Risk:
+        It might allow other AWS identities or the root access of the account to escalate privileges.
+
+    Fix:
+        Where possible, restrict the access to only the required resources.
+        For example, instead of `Principal: "*"`, include a list of the roles that need access.
+
+    Filters context:
+        | Parameter   | Type               | Description                                                    |
+        |:-----------:|:------------------:|:--------------------------------------------------------------:|
+        |`config`     | `str`              | `config` variable available inside the rule                    |
+        |`extras`     | `str`              | `extras` variable available inside the rule                    |
+        |`logical_id` | `str`              | ID used in CloudFormation to refer the resource being analysed |
+        |`resource`   | `S3BucketPolicy`   | Resource that is being addressed                               |
+        |`statement`  | `Statement`        | Statement being checked found in the Resource                  |
+        |`principal`  | `str`              | AWS Principal being checked found in the statement             |
+        |`account_id` | `str`              | Account ID found in the principal                              |
+    """
+
+    REASON_WILDCARD_PRINCIPAL = (
+        "{} should not allow wildcard in principals or account-wide principals (principal: '{}')"
+    )
+    RISK_VALUE = RuleRisk.MEDIUM
+    FULL_REGEX = REGEX_PARTIAL_WILDCARD_PRINCIPAL
+
+
+class GenericResourceFullWildcardPrincipalRule(GenericResourceWildcardPrincipalRule):
     """
     Checks for any wildcard principal defined in any statement.