+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.6.0]
+### Updates
+- Created `GenericResourceWildcardPrincipalRule` to be an abstract for wildcard principals for Generic resources.
+- Created `GenericResourcePartialWildcardPrincipalRule` and `GenericResourceFullWildcardPrincipalRule` to evaluate Generic resources.
+### Fixes
+- Rollback `GenericWildcardPrincipalRule` as it was in `1.5.2`.
+
 ## [1.5.3]
 ### Updates
 - Updates `GenericWildcardPrincipalRule` to understand the `GenericResource`.